feat: initial setup of the details
