+++ src/Components/Base/BaseCarousel/FullscreenBaseCarousel.tsx
@@ -26,6 +26,7 @@ type Props = {
     | "testID"
     | "itemHeight"
     | "gap"
+    | "rootStyle"
 >
 
 export const FullscreenBaseCarousel = ({
