+++ src/Hooks/useTransactionScreen/useTransactionScreen.tsx
@@ -386,5 +386,6 @@ export const useTransactionScreen = ({
         isFirstTimeLoadingFees,
         hasEnoughBalanceOnToken,
         isBiometricsEmpty,
+        transactionOutputs,
     }
 }
